requirements.txt
asgiref==3.7.2
Django==3.2
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
psycopg2-binary==2.9.1
PyJWT==1.7.1
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
